Added to_h method to Offense
The Offense base class implements a to_h method with the offense details, this way we can use the resulting hash for multipourposes.
Also, the offense is always initialized with an uncorrected status, the status has an attribute writter to change it to corrected when the client runs with the --autocorrect  option and has a new method to check if the offense was corrected or not.