Stop erasing items in ARGV
When running erb-lint, ARGV was being modified and all its items were
being removed. This was causing issues when printing error messages
with the original command that was run by users. This is happening
because the code is modifying an object being passed by reference.
Instead, this dups ARGV so it stays intact.