Added Rake task
Necessary to provide a Rake task which can be configured with custom behavior if the defaults are not desired.

During code review, it was requested that `Rake#rake_output_message` be used but this is a wrapper around `$stderr` and this message should really be sent to `$stdout` instead since it's not, technically, and error. Alas, the change was made anyway.