Centralize Ruby Version to `.ruby-version` (#345)
* Update rubocop to 1.62.1

* Centralize Ruby Version to `.ruby-version`

The `.ruby-version` file is the ecosystem standard for defining a Ruby version.  This PR adds the `.ruby-version` file, ensures a `required_ruby_version` is set, and removes all other references to Ruby in this repository, aligning it with the standard.
> [!IMPORTANT]
> Please verify the following before merging:

Verify that the changes in the PR meets the following requirements or adjust manually to make it compliant:
  - [ ] `.ruby-version` file is present with the correct Ruby version defined
  - [ ] A `required_ruby_version` in your gemspec is set
  - [ ] There is no Ruby version present in the  `dev.yml` Ruby task (before: `- ruby: x.x.x`, after: `- ruby`)
  - [ ] There is no Ruby version/requirement referenced in the `Gemfile` (no lines with `ruby  <some-version>`)
  - [ ] A `Gemfile.lock` is built with the defined Ruby version
  - [ ] The version of Rubocop installed is 1.61.0 or greater
  - [ ] There is no `TargetRubyVersion` defined  in `rubocop.yml`
  - [ ] There is no Ruby argument  present in  `ruby/setup-ruby` Github Actions that do **not**  run on a Ruby matrix (no lines with `ruby-version: “x.x”`)

This PR will be merged if there isn't any activity after 4 weeks.

* Temporarily rename the .ruby-version file to pass CI for rubocop-old

Old versions of rubocop read the ruby version from .ruby-version first.
We made the change upstream which is not captured in versions of
rubocop (< 0.87).

Upstream commit: https://github.com/rubocop/rubocop/commit/02b2c3a461f197134cb1d7e1bf693c4e3902e8de

Co-authored-by: Étienne Barrié <etienne.barrie@shopify.com>

* Remove ruby-version mention from linting.yml

* Update rubocop-shopify

* Disable Style/ClassMethodsDefinitions

* Run bundle exec rubocop -a

Ignore the .intersect? autocorrect

* Update better_html to 2.1.1

* Fix consistency typo in CI test versions

---------

Co-authored-by: Étienne Barrié <etienne.barrie@shopify.com>